+++ src/main/java/com/fpt/capstone/tourism/repository/RoleRepository.java
@@ -14,6 +14,6 @@ public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByRoleName(String roleName);
     Optional<Role> findById(Long roleId);
     @Query("SELECT r FROM Role r JOIN UserRole ur ON r.id = ur.role.id WHERE ur.user.id = :userId")
-    List<Role> findRolesByUserId(@Param("userId") int userId);
+    List<Role> findRolesByUserId(@Param("userId") Long userId);
 }
